#147 - tidy graphQL schema and place related objects together
