+++ api/src/schemas/game.js
@@ -11,6 +11,19 @@ import sanitiseHtml from 'utils/sanitiseHtml';
 import serviceExecutor from 'utils/serviceExecutor';
 
 export const gameTypeDefs = `
+
+  extend type Query {
+    game(id: ID!): Game!
+    games(offset: Int, searchOptions: SearchOptions): [Game]
+    myGames(status: [String]): [Game]
+  }
+  
+  extend type Mutation {
+    createGame(input: CreateGameInput): Game
+    updateGame(id: ID!, input: UpdateGameInput): Game
+    updateGameStatus(id: ID!, gameStatus: Int): Game
+  }
+
   type GameSetting {
     minPlayers: Int!,
     maxPlayers: Int!,
