+++ api/src/schemas/gamePlayer.js
@@ -11,6 +11,23 @@ export const GAME_PLAYER_UPDATED = 'game_player_updated';
 
 export const gamePlayerTypeDefs = `
 
+  extend type Query {
+    gamePlayer(id: ID!): GamePlayer!,
+    myGamePlayer(gameId: ID!): [GamePlayer]
+    myGamePlayers: [GamePlayer]
+    gamePlayers(gameId: ID!, status: [String]): [GamePlayer!]
+  }
+  
+  extend type Mutation {
+    createGamePlayer(input: CreateGamePlayerInput): GamePlayer
+    updateGamePlayer(id: ID!, input: UpdateGamePlayerInput): GamePlayer
+  }
+  
+  extend type Subscription {
+    gamePlayerAdded(gameId: ID!): GamePlayer!
+    gamePlayerUpdated(gameId: ID!): GamePlayer!
+  }
+
   type GamePlayer {
     id: ID!,
     user: User!,
